refactor: move TapListener to services package for shared usage
The TapListener class was previously located in the godmode feature package but is used by multiple features (godmode, smartdrill, smartupgrade). Moving it to a common services package improves code organization and reusability across the codebase.
