+++ src/mindustrytool/Main.java
@@ -29,11 +29,11 @@ import mindustrytool.features.settings.FeatureSettingDialog;
 import mindustrytool.features.smartupgrade.SmartUpgradeFeature;
 import mindustrytool.features.smartdrill.SmartDrillFeature;
 import mindustrytool.services.ServerService;
+import mindustrytool.services.TapListener;
 import mindustrytool.services.CrashReportService;
 import mindustrytool.services.UpdateService;
 import mindustrytool.features.chat.global.ChatFeature;
 import mindustrytool.features.godmode.GodModeFeature;
-import mindustrytool.features.godmode.TapListener;
 import mindustrytool.features.autoplay.AutoplayFeature;
 import mindustrytool.features.background.BackgroundFeature;
 import mindustrytool.features.music.MusicFeature;
